split out a separate entrypoint for the bloch sphere widget
