+++ source/widgets/js/index.tsx
@@ -18,7 +18,6 @@ import {
   MoleculeViewer,
   Entanglement,
   type EntanglementProps,
-  BlochSphere,
 } from "qsharp-lang/ux";
 import markdownIt from "markdown-it";
 import "./widgets.css";
@@ -91,9 +90,6 @@ function render({ model, el }: RenderArgs) {
       break;
     case "Entanglement":
       renderEntanglement({ model, el });
-      break;
-    case "BlochSphere":
-      renderBlochSphere({ model, el });
       break;
     default:
       throw new Error(`Unknown component type ${componentType}`);
@@ -291,25 +287,6 @@ function renderCircuit({ model, el }: RenderArgs) {
 
   onChange();
   model.on("change:circuit_json", onChange);
-}
-
-function renderBlochSphere({ model, el }: RenderArgs) {
-  const onChange = () => {
-    const initialGates = model.get("initial_gates") as string;
-    prender(
-      <BlochSphere
-        initialGates={initialGates}
-        onGatesChanged={(gates) => {
-          model.set("gates", gates);
-          model.save_changes();
-        }}
-      ></BlochSphere>,
-      el,
-    );
-  };
-
-  onChange();
-  model.on("change:initial_gates", onChange);
 }
 
 function renderAtoms({ model, el }: RenderArgs) {
