+++ source/widgets/package.json
@@ -2,7 +2,7 @@
   "scripts": {
     "dev": "npm run build:bundle -- --sourcemap=inline --watch",
     "build": "npm run build:bundle && node js/subset-katex-fonts.mjs",
-    "build:bundle": "npx esbuild js/index.tsx js/bloch.css --minify --format=esm --bundle --outdir=src/qsharp_widgets/static --loader:.woff2=dataurl --loader:.woff=empty --loader:.ttf=empty"
+    "build:bundle": "npx esbuild js/index.tsx js/bloch.tsx --minify --format=esm --bundle --outdir=src/qsharp_widgets/static --loader:.woff2=dataurl --loader:.woff=empty --loader:.ttf=empty"
   },
   "devDependencies": {}
 }
