+++ source/widgets/src/qsharp_widgets/__init__.py
@@ -213,7 +213,7 @@ class Circuit(anywidget.AnyWidget):
 
 
 class BlochSphere(anywidget.AnyWidget):
-    _esm = pathlib.Path(__file__).parent / "static" / "index.js"
+    _esm = pathlib.Path(__file__).parent / "static" / "bloch.js"
     _css = pathlib.Path(__file__).parent / "static" / "bloch.css"
 
     comp = traitlets.Unicode("BlochSphere").tag(sync=True)
